requirements.txt
streamlit==1.22.0
pandas==1.5.3
numpy==1.23.5
mysql-connector==2.2.9
ydata-profiling==4.1.2
streamlit-player==0.1.5
streamlit-pandas-profiling==0.1.3
streamlit-extras==0.2.7
plotly==5.14.1
seaborn==0.12.2
git-python==1.0.3
XlsxWriter==3.1.0